Move abstract view to qr module
