+++ collect_app/src/main/res/layout/qr_code_project_creator_dialog_layout.xml
@@ -39,7 +39,7 @@
                     app:layout_constraintStart_toStartOf="parent"
                     app:layout_constraintTop_toTopOf="parent" />
 
-                <org.odk.collect.android.fragments.BarcodeScannerViewContainer
+                <org.odk.collect.qrcode.BarcodeScannerViewContainer
                     android:id="@+id/barcode_view"
                     app:layout_constraintWidth_percent="0.75"
                     app:layout_constraintDimensionRatio="w,1:1"
